refactor(database_utils): another one bites the dust
