+++ crates/utils/database/src/lib.rs
@@ -16,11 +16,17 @@ use fitness_models::UserMeasurementsListInput;
 use itertools::Itertools;
 use markdown::to_html as markdown_to_html;
 use media_models::{ImportOrExportItemRating, ImportOrExportItemReview, ReviewItem};
+use migrations::AliasedCollectionToEntity;
 use rust_decimal_macros::dec;
 use sea_orm::{
-    ColumnTrait, DatabaseConnection, EntityTrait, ModelTrait, QueryFilter, QueryOrder, QueryTrait,
+    prelude::Expr, sea_query::PgFunc, ColumnTrait, DatabaseConnection, EntityTrait, ModelTrait,
+    QueryFilter, QueryOrder, QuerySelect, QueryTrait, Select,
 };
 use user_models::UserReviewScale;
+
+pub fn ilike_sql(value: &str) -> String {
+    format!("%{value}%")
+}
 
 pub async fn user_by_id(db: &DatabaseConnection, user_id: &String) -> Result<user::Model> {
     User::find_by_id(user_id)
@@ -187,4 +193,45 @@ pub async fn workout_details(
             })
         }
     }
+}
+
+pub fn apply_collection_filter<E, C, D>(
+    query: Select<E>,
+    collection_id: Option<Vec<String>>,
+    invert_collection: Option<bool>,
+    entity_column: C,
+    id_column: D,
+) -> Select<E>
+where
+    E: EntityTrait,
+    C: ColumnTrait,
+    D: ColumnTrait,
+{
+    query.apply_if(collection_id, |query, v| {
+        let unique_collections = v.into_iter().unique().collect_vec();
+        let count = unique_collections.len() as i32;
+        let subquery = CollectionToEntity::find()
+            .select_only()
+            .column(id_column)
+            .filter(
+                Expr::col((
+                    AliasedCollectionToEntity::Table,
+                    collection_to_entity::Column::CollectionId,
+                ))
+                .eq(PgFunc::any(unique_collections)),
+            )
+            .filter(id_column.is_not_null())
+            .group_by(id_column)
+            .having(
+                collection_to_entity::Column::CollectionId
+                    .count()
+                    .eq(Expr::val(count)),
+            )
+            .into_query();
+        if invert_collection.unwrap_or_default() {
+            query.filter(entity_column.not_in_subquery(subquery))
+        } else {
+            query.filter(entity_column.in_subquery(subquery))
+        }
+    })
 }
